update page titles in Routes, change NewReport to UploadReport
