--- tdrs-frontend/src/components/NewReport/index.js
@@ -1,3 +1,0 @@
-import NewReport from './NewReport'
-
-export default NewReport
