+++ tdrs-frontend/src/components/Reports/Reports.jsx
@@ -27,39 +27,34 @@ function Reports() {
     return options
   }
   return (
-    <div className="grid-container">
-      <h1 className="font-serif-2xl margin-bottom-0 text-normal">
-        All Reports
-      </h1>
-      <form className="usa-form">
-        {/* eslint-disable-next-line */}
+    <form className="usa-form">
+      {/* eslint-disable-next-line */}
         <label className="usa-label text-bold margin-top-4" htmlFor="reports">
-          Reporting Year
-        </label>
-        {/* eslint-disable-next-line */}
+        Reporting Year
+      </label>
+      {/* eslint-disable-next-line */}
         <select
-          className="usa-select"
-          name="options"
-          id="reports"
-          onChange={handleSelect}
-        >
-          {createOptions(3)}
-        </select>
+        className="usa-select"
+        name="options"
+        id="reports"
+        onChange={handleSelect}
+      >
+        {createOptions(3)}
+      </select>
 
-        <h2 className="font-serif-xl margin-y-4 text-normal">
-          {selectedYear} TANF Reports
-        </h2>
+      <h2 className="font-serif-xl margin-y-4 text-normal">
+        {selectedYear} TANF Reports
+      </h2>
 
-        <p className="font-sans-md margin-0 text-bold">Test Quarter</p>
-        <p className="font-sans-md margin-0 margin-bottom-105">
-          Upload by a date
-        </p>
+      <p className="font-sans-md margin-0 text-bold">Test Quarter</p>
+      <p className="font-sans-md margin-0 margin-bottom-105">
+        Upload by a date
+      </p>
 
-        <Button type="button" onClick={handleClick}>
-          Begin Report
-        </Button>
-      </form>
-    </div>
+      <Button type="button" onClick={handleClick}>
+        Begin Report
+      </Button>
+    </form>
   )
 }
