+++ tdrs-frontend/src/components/UploadReport/index.js
@@ -1,0 +1,3 @@
+import UploadReport from './UploadReport'
+
+export default UploadReport
